Deck between Scenes + card used displa + animations
created the deck as a child of the player and then in the battle scene it creates a clone and uses the clone to run the battle.
Created animations for the block and attack.
Made the discard pile look good and draw
Made a cards used display and displays the used cards and you cannot use more than the 2 cards given.