added next move icon, win and lose
Added a win canvas that moves onto the screen when you when and pauses the game, and makes the end turn button unaccessible. Made a similar lose canvas.

Added an image and text that updates based on how much damage the enemy does. It will show the next move and the amount for the next move.