Game now saves
Saves on encounter completion and on hitting the escape key.
IMPORTANT: set an ID to all Cards through a serializable value on the card script based on its position on the Cards/Cards folder. This is used for the first save and is very important.

Things Saved (in player prefs):
-player health and satiation
-the ids of the cards in the deck
-whether each encounter has been completed (saved in key of encounterName)
-the player's position